Lens/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
numpy
sqlalchemy
databases
python-dotenv
aiohttp
httpx
scikit-learn
matplotlib
seaborn
jinja2
pydantic
openai
sqlalchemy-utils
psycopg2-binary
pytest
